+++ src/main/java/com/pixel/synchronre/sychronremodule/model/entities/Tranche.java
@@ -28,7 +28,7 @@ public class Tranche
     private BigDecimal tranchePriorite;
     private BigDecimal tranchePorte;
     private BigDecimal trancheTauxPrime;
-    private int trancheNumero;
+    private Long trancheNumero;
     @ManyToOne @JoinColumn(name = "risque_couvert_ID")
     private RisqueCouvert risqueCouvert;
